pyproject.toml
[project]
name = "django-template"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.2.1",
    "pre-commit>=4.2.0",
    "python-decouple>=3.8",
    "pyyaml>=6.0.2",
    "ruff>=0.11.13",
    "uvicorn>=0.34.2",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.format]
quote-style = "single"
line-ending = "lf"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "N", "SIM", "UP"]

[tool.ruff.lint.isort]
combine-as-imports = true
force-single-line = false
